# REST-api/openapi.et.yaml
openapi: 3.0.0
info:
  title: Google Forms klooni API
  version: "1.0.0"
  description: >
    Google Forms sarnase teenuse OpenAPI näidisspetsifikatsioon.
    Kasutab ressursside nimedena üksikuid mitmuse nimisõnu ja /sessions autentimiseks.
servers:
  - url: /
    description: Tootmisserver
...
paths:
...
  /users:
    get:
      tags:
        - users
      summary: Kuva kõik kasutajad
      operationId: listUsers
      responses:
        "200":
          description: Kasutajate nimekiri
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Autoriseerimata
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]

    post:
      tags:
        - users
      summary: Loo uus kasutaja (registreeru)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: Kasutaja on loodud
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Päringus on puudu vajalikud väljad
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Päringus on puudu vajalikud väljad: email, parool ja nimi on nõutud. Parool peab olema vähemalt 8 tähemärki pikk, Parool peab sisaldama vähemalt 3 järgnevast: suurtähed, väiketähed, numbrid, erimärgid"
        "409":
            description: Selle emailiga kasutaja on juba olemas
        "500":
          $ref: "#/components/responses/ServerError"
  /users/{userId}:
    get:
      tags:
        - users
      summary: Kuva konkreetse kasutaja üksikasjad
      operationId: getUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Kasutaja üksikasjad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Kasutajat ei leitud
        "401":
          description: Autoriseerimata
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]
    patch:
      tags:
        - users
      summary: Uuenda olemasoleva kasutaja osa
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Kasutaja on uuendatud
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Kasutajat ei leitud
        "401":
          description: Autoriseerimata
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]
    delete:
      tags:
        - users
      summary: Kustuta kasutaja
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Kasutaja on kustutatud
        "404":
          description: Kasutajat ei leitud
        "401":
          description: Autoriseerimata
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Vigane päring - vigased sisendparameetrid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Serveri sisemine viga
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - message
...
    UserCreate:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        name:
          type: string
      required:
        - email
        - password
        - name
    UserUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        name:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        passwordUpdated:
          type: boolean
          description: Kas parool on uuendatud pärast viimast sisselogimist
      required:
        - id
        - email
        - name
        - createdAt
        - updatedAt
        - passwordUpdated